chore: replace proxy.env by template
If proxy settings need to be configured, the proxy.env file is always
reported by git as modified.  To prevent this, it has been replaced by a
template file. The template can be copied to proxy.env and modified.
proxy.env has been added to .gitignore; therefore it can be freely
manipulated.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>