chore(v2): add shell script to safely start up SW360
There is a race condition between starting the CouchDB and the Thrift
container; the database is not yet available when the Thrift container
comes up. The synthetic start_dependencies service fixes this, but it
has to be executed using a separate docker-compose command. To
simplify these steps, a script to startup the whole system is added.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>