chore(v2): get couchdb-lucene integration working
The Lucene query functionality works differently with CouchDB 2.
Queries now have to be sent directly to the couchdb-lucene service. To
enable this, the couchdb.properties properties file now contains the
URI of this service. It is passed to the Thrift service as an
environment variable to replace it in the couchdb.properties.template
file.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>