chore(v2): enable building of images via docker-compose
Some images depend on an SW360 revision. The compose file now passes
this revision as argument to the docker images. The default value is
defined in an .env file, but it can be overridden by specifying a
--build-arg on the command line.

Note that the default revision is currently not an official tag. This
is because the latest 11.0.0-M1 tag has a problem with searches via
Lucene. So the commit is referenced, which fixes this problem.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>