chore(v2): automate CouchDB 2 setup
CouchDB requires a manual setup step to configure whether it should
run as single or clustered node; in addition, some system databases
need to be created. If this is not done, errors are logged
continuously.

The entrypoint script of the backend container now checks whether
system tables are present. If not, the one-time initialization is
done.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>